+++ src/burn/drv/pre90s/d_tecmo.cpp
@@ -685,7 +685,7 @@ static void TecmoFMIRQHandler(INT32, INT32 nStatus)
 
 static INT32 TecmoSynchroniseStream(INT32 nSoundRate)
 {
-	return (INT64)ZetTotalCycles() * nSoundRate / 4000000;
+	return (INT64)(double)ZetTotalCycles() * nSoundRate / 4000000;
 }
 
 static void TecmoMSM5205Vck()
