+++ src/burn/drv/pre90s/d_toki.cpp
@@ -906,7 +906,7 @@ static void DrvFMIRQHandler(INT32, INT32 nStatus)
 
 inline static INT32 TokibSynchroniseStream(INT32 nSoundRate)
 {
-	return (INT64)(ZetTotalCycles() * nSoundRate / 4000000);
+	return (INT64)((double)ZetTotalCycles() * nSoundRate / 4000000);
 }
 
 static void toki_adpcm_int()
